+++ R/pick_phylopic.R
@@ -2,7 +2,7 @@
 #'
 #' This function provides a visually interactive way to pick an image and
 #' valid uuid for an input taxonomic name. As multiple silhouettes can exist
-#' for each species in PhyloPic, this function is useful for choosing the
+#' for each organism in PhyloPic, this function is useful for choosing the
 #' right image/uuid for the user.
 #'
 #' @param name \code{character}. A taxonomic name. Different taxonomic levels
@@ -11,10 +11,16 @@
 #'   requested `name`, multiple silhouettes may exist. If `n` exceeds the
 #'   number of available images, all available uuids will be returned.
 #'   Defaults to 5.
+#' @param ncol \code{numeric}. Number of columns in the plot grid (default:
+#'   1). The number of silhouettes plotted at once is the product of `ncol`
+#'   and `nrow`.
+#' @param nrow \code{numeric}. Number of rows in the plot grid (default: 1).
+#'   The number of silhouettes plotted at once is the product of `ncol` and
+#'   `nrow`.
 #' @param auto \code{numeric}. This argument allows the user to automate input
-#'   into the menu choice. If the input value is `1`, requested images will be
-#'   automatically cycled through with the final image returned. If the input
-#'   value is `2`, the first returned image will be selected. If `NULL`
+#'   into the menu choice. If the input value is `1`, the first returned image 
+#'   will be selected. If the input value is `2`, requested images will be
+#'   automatically cycled through with the final image returned. If `NULL`
 #'   (default), the user must interactively respond to the called menu.
 #'
 #' @return A [Picture][grImport2::Picture-class] object is returned. The uuid
@@ -25,28 +31,31 @@
 #'   from a pool of silhouettes available for the input `name`.
 #'
 #' @importFrom grid grid.newpage grid.text
-#' @importFrom grImport2 grid.picture
+#' @importFrom grImport2 grid.picture grobify
 #' @importFrom utils menu
+#' @importFrom ggpubr ggarrange
 #' @export
 #' @examples \dontrun{
+#' # Defaults pane layout
 #' img <- pick_phylopic(name = "Canis lupus", n = 5)
+#' # 3 x 3 pane layout
+#' img <- pick_phylopic(name = "Scleractinia", n = 9, ncol = 3, nrow = 3)
 #' }
-pick_phylopic <- function(name = NULL, n = 5, auto = NULL) {
+pick_phylopic <- function(name = NULL, n = 5, ncol = 1, nrow = 1, auto = NULL) {
   # Error handling
   if (!is.null(auto) && !auto %in% c(1, 2)) {
     stop("`auto` must be of value: NULL, 1, or 2")
   }
-  # Get uuids
-  uuids <- get_uuid(name = name, n = n, url = FALSE)
-  # Cycle through uuids
-  for (i in seq_along(uuids)) {
-    # Start a new graphics page
+  if (!is.numeric(ncol) || !is.numeric(nrow)) {
+    stop("`ncol` and `nrow` must be of class numeric.")
+  }
+  
+  # Internal function for plotting selected image
+  return_img <- function(uuids) {
+    img <- get_phylopic(uuid = uuids)
+    att <- get_attribution(uuid = uuids)
+    print(uuids)
     grid.newpage()
-    # Get image data
-    img <- get_phylopic(uuid = uuids[i], format = "vector")
-    # Get attribution data
-    att <- get_attribution(uuid = uuids[i])
-    # Plot image
     grid.picture(img)
     # Add text for attribution
     att_string <- paste0("Contributor: ", att$contributor, "\n",
@@ -56,22 +65,86 @@ pick_phylopic <- function(name = NULL, n = 5, auto = NULL) {
               x = 0.96, y = 0.92,
               just = "right",
               gp = gpar(fontsize = 8, col = "purple", fontface = "bold"))
-    # Return data for final image
-    if (i == length(uuids)) {
-      message("This is the only or final image. Returning this uuid data.")
-      print(uuids[i])
+    return(img)
+  }
+
+  # Get uuids
+  uuids <- get_uuid(name = name, n = n, url = FALSE)
+  # Record length
+  n_uuids <- length(uuids)
+  
+  # Return data if only one image requested
+  if (n == 1) {
+    img <- return_img(uuids = uuids)
+    return(img)
+  }
+  
+  # Return data if only one image
+  if (n_uuids == 1) {
+    message("This is the only image. Returning this uuid data.")
+    img <- return_img(uuids = uuids)
+    return(img)
+  }
+  
+  # Generate list of uuids for batch viewing
+  # How many silhouettes should be plotted at once?
+  batch <- prod(ncol, nrow)
+  
+  # Suppress warnings when there is an uneven split
+  if ((length(uuids) %% batch) != 0) {
+    uuids <- suppressWarnings(
+      split(x = uuids, f = ceiling(seq_along(uuids) / batch)))
+  } else {
+    uuids <- split(x = uuids, f = ceiling(seq_along(uuids) / batch))
+  }
+  
+  # Cycle through list
+  for (i in seq_along(uuids)) {
+    # Get image data
+    img <- sapply(uuids[[i]], get_phylopic)
+    # Get attribution data
+    att <- lapply(uuids[[i]], get_attribution)
+    # Attribution text
+    att_string <- lapply(att, function(x){
+      paste0(x$contributor, " (", x$created, "). License: ", x$license)
+    })
+    att_string <- unlist(att_string)
+    
+    # Set up menu
+    if (is.null(auto)) {
+      # Grobify images
+      img <- lapply(img, grobify)
+      # Plot image choices
+      print(ggarrange(plotlist = img, nrow = nrow, ncol = ncol, 
+                      labels = 1:length(img)))
+      m <- menu(choices = c(att_string, "Next"), title = paste0(
+        "Choose an option (", i, "/", ceiling(n_uuids / batch), " pages):"))
+    } else {
+      # Select final uuid
+      if (auto == 2) {
+        # Update i (final batch)
+        i <- length(uuids)
+        # Update m  to 'next' valye (force final image of final batch)
+        n_plotted <- length(uuids[[i]])
+        m <- n_plotted + 1
+      } else if (auto == 1) {
+        m <- 1
+      }
+    }
+    
+    # Make selection
+    n_plotted <- length(uuids[[i]])
+    if (m != (n_plotted + 1)) {
+      uuids <- uuids[[i]][m]
+      img <- return_img(uuids = uuids)
       return(img)
     }
-    # Set up menu choice
-    if (is.null(auto)) {
-      m <- menu(choices = c("Next", "Select"), title = paste0(
-        "Choose an option (", i, "/", length(uuids), "):"))
-    } else {
-      m <- auto
-    }
-    # Make selection
-    if (m == 2) {
-      print(uuids[i])
+    
+    # If final image available reached, return
+    if (i == length(uuids)) {
+      message("This is the final image. Returning this uuid data.")
+      uuids <- uuids[[i]][n_plotted]
+      img <- return_img(uuids = uuids)
       return(img)
     }
   }